requirements.txt
Flask
mysql-connector-python
pytesseract
Pillow